respond to code review comments
