+++ deepcell_applications/utils.py
@@ -25,8 +25,7 @@
 # ==============================================================================
 """Functions for instantiating and running Applications"""
 
-import argparse
-import os
+
 
 import deepcell_applications as dca
 
@@ -93,98 +92,3 @@ def get_predict_kwargs(kwargs):
     return predict_kwargs
 
 
-def get_arg_parser():
-    """argument parser to consume command line arguments"""
-    root_dir = os.path.dirname(os.path.abspath(__file__))
-    parser = argparse.ArgumentParser()
-
-    # Create a parent parser for common arguments
-    # https://stackoverflow.com/a/63283912
-    parent = argparse.ArgumentParser()
-
-    class WritableDirectoryAction(argparse.Action):
-        # Check that the provided output directory exists, and is writable
-        # From: https://gist.github.com/Tatsh/cc7e7217ae21745eb181
-        def __call__(self, parser, namespace, values, option_string=None):
-            prospective_dir = values
-            if not os.path.isdir(prospective_dir):
-                raise argparse.ArgumentTypeError(
-                    '{} is not a valid directory'.format(
-                        prospective_dir, ))
-            if os.access(prospective_dir, os.W_OK | os.X_OK):
-                setattr(namespace, self.dest, os.path.realpath(prospective_dir))
-                return
-            raise argparse.ArgumentTypeError(
-                '{} is not a writable directory'.format(
-                    prospective_dir))
-
-    def existing_file(x):
-        if x is not None and not os.path.exists(x):
-            raise argparse.ArgumentTypeError('{} does not exist.'.format(x))
-        return x
-
-    parent.add_argument('--output-directory', '-o',
-                        default=os.path.join(root_dir, 'output'),
-                        action=WritableDirectoryAction,
-                        help='Directory where application outputs are saved.')
-
-    parent.add_argument('--output-name', '-f',
-                        default='mask.tif',
-                        help='Name of output file.')
-
-    parent.add_argument('-L', '--log-level', default='INFO',
-                        choices=('DEBUG', 'INFO', 'WARN', 'ERROR', 'CRITICAL'),
-                        help='Only log the given level and above.')
-
-    parent.add_argument('--squeeze', action='store_true',
-                        help='Squeeze the output tensor before saving.')
-
-    # use subparsers to group options for different applications
-    # https://stackoverflow.com/a/30217387
-    subparsers = parser.add_subparsers(dest='app', help='application name')
-
-    # Next, each application is configured as its own subparser
-    # Each subparser should inherit from ``parent`` to inherit options
-    # Configurable inputs for ``prepare_input`` and ``app.predict`` must
-    # match (via name or dest) the function input names exactly.
-
-    # Mesmer Application Configuration
-    mesmer = subparsers.add_parser('mesmer', parents=[parent], add_help=False,
-                                   help='Run Mesmer on nuclear + membrane data')
-
-    # Mesmer Image file inputs
-    mesmer.add_argument('--nuclear-image', '-n', required=True,
-                        type=existing_file, dest='nuclear_path',
-                        help=('REQUIRED: Path to 2D single channel TIF file.'))
-
-    mesmer.add_argument('--nuclear-channel', '-nc',
-                        default=0, nargs='+', type=int,
-                        help='Channel(s) to use of the nuclear image. '
-                             'If more than one channel is passed, '
-                             'all channels will be summed.')
-
-    mesmer.add_argument('--membrane-image', '-m',
-                        type=existing_file, dest='membrane_path',
-                        help=('Path to 2D single channel TIF file. '
-                              'Optional. If not provided, membrane '
-                              'channel input to network is blank.'))
-
-    mesmer.add_argument('--membrane-channel', '-mc',
-                        default=0, nargs='+', type=int,
-                        help='Channel(s) to use of the membrane image. '
-                             'If more than one channel is passed, '
-                             'all channels will be summed.')
-
-    # Mesmer Inference parameters
-    mesmer.add_argument('--image-mpp', type=float, default=0.5,
-                        help='Input image resolution in microns-per-pixel. '
-                             'Default value of 0.5 corresponds to a 20x zoom.')
-
-    mesmer.add_argument('--batch-size', '-b', default=4, type=int,
-                        help='Batch size for `model.predict`.')
-
-    mesmer.add_argument('--compartment', '-c', default='whole-cell',
-                        choices=('nuclear', 'membrane', 'whole-cell'),
-                        help=('The cellular compartment to segment.'))
-
-    return parser
